+++ tap/Makefile
@@ -1,2 +1,5 @@
 run:
 	python tap.py
+
+install:
+	pip install -r requirements.txt
